Decompiles Various `bo4` and `rbo5` Functions
Decompiles several duplicates across `bo4` and `rbo5` and several also
shared with `dra`.

Notably, the function signature for `g_api.func_800EB758` seems to
differ from the actual `func_800EB758` function signature. This function
pointer is only used (so far) by `bo4` and `rbo5`.

+++ include/game.h
@@ -1579,8 +1579,10 @@ typedef struct {
     /* 8003C7E4 */ void (*func_800EA538)(s32 arg0);
     /* 8003C7E8 */ void (*g_pfn_800EA5AC)(u16 arg0, u8 arg1, u8 arg2, u8 arg3);
     /* 8003C7EC */ void (*func_801027C4)(u32 arg0);
+    // this signature differs from `func_800EB758`. the last
+    // argument is 16-bits instead of 8.
     /* 8003C7F0 */ void (*func_800EB758)(
-        s16 pivotX, s16 pivotY, Entity* e, u16 flags, POLY_GT4* p, u8 flipX);
+        s16 pivotX, s16 pivotY, Entity* e, u16 flags, POLY_GT4* p, u16 flipX);
     /* 8003C7F4 */ Entity* (*CreateEntFactoryFromEntity)(
         Entity* self, u32 flags, s32 arg2);
     /* 8003C7F8 */ bool (*func_80131F68)(void);
